adding files

+++ controllers/admin.js
@@ -1,7 +1,8 @@
 //const mongodb = require('mongodb');
-//const mongoose = require('mongoose');
-const { validationResult } = require('express-validator/check');
+const mongoose = require('mongoose');
+const { validationResult } = require('express-validator');// /check
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 //const ObjectId = mongodb.ObjectId;
 
@@ -22,9 +23,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  //const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  console.log(image);
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+      //isAuthenticated : req.session.isLoggedIn
+    });
+  }
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -45,6 +64,7 @@ exports.postAddProduct = (req, res, next) => {
       //isAuthenticated : req.session.isLoggedIn
     });
   }
+  const imageUrl = image.path;
 
   const product = new Product({
     // _id: new mongoose.Types.ObjectId('609b36fe6b576700045f16f6'),
@@ -128,10 +148,12 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  //const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
+
   if (!errors.isEmpty()) {
     console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
@@ -141,7 +163,7 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
+        //imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -162,7 +184,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      //product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
 
       return product.save()
         .then(result => {
@@ -202,7 +228,18 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
+    // .catch(err => next(err));
+
+
     //Product.findByIdAndRemove(prodId)
     .then(() => {
       console.log('DESTROYED PRODUCT');
